+++ ReleaseNotes.md
@@ -1,6 +1,11 @@
 # Release notes of Ampersand
 
-## unreleased changes
+## Unreleased changes
+  * [Issue #600](https://github.com/AmpersandTarski/Ampersand/issues/600) When a configuration file exists (.yaml) with the same name as the file being compiled, it will be used if no config file is explicitly mentioned in the command. 
+
+## v3.8.0 (20 december 2016) 
+  * [Issue #587](https://github.com/AmpersandTarski/Ampersand/issues/587) There is no need any longer to explicitly specify the language in a script. If omitted, IN ENGLISH will be used as default. 
+  * [Issue #588](https://github.com/AmpersandTarski/Ampersand/issues/588) PURPOSE sytax changed: Now matching brackets are: `{+` and `+}`   (no backwards compatibility, to enable nested comment blocks)
   * Added some additional diagnosis info in settings.json
   * Added support for project/application and extension specific composer dependencies.
   * Added 'customization' folder to prototype generation process. This folder can be used to e.g. overwrite generated views.